requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
pandas
anthropic
weasyprint
markdown
python-jobspy
